geo entity rendering (textureless)

+++ ulivy_shaders/basic_entity_fs.glsl
@@ -7,9 +7,9 @@ precision highp float;
 out vec4 frag_color;
 
 out vec2 fSize;
-out ivec2 fTexPos;
-out ivec2 fTexSize;
-out ivec2 fTexFrame;
+out vec2 fTexPos;
+out vec2 fTexSize;
+out vec2 fTexFrame;
 
 void main()
 {
